requirements.txt
google-generativeai
python-dotenv
streamlit
pandas
numpy
matplotlib
scikit-learn
Pillow
